requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
wincertstore==0.2
